+++ .gitignore
@@ -66,3 +66,4 @@ next-env.d.ts
 packages/create-message-kit/templates/**
 packages/create-message-kit/templates/**/*
 packages/create-message-kit/templates
+templates/toss/**/*
