+++ packages/docs/pages/concepts/structure.mdx
@@ -43,7 +43,6 @@ run(
 - `experimental`: experimental features like logging all group messages. Default is `false`.
 - `attachments`: to receive attachments. Default is `false`.
 - `gptModel`: model to be used. Default is `gpt-4o`.
-- `walletService`: if true, the agent cdp wallet will be used. Default is `false`.
 - `client`: Optional parameters to pass to the XMTP client.
 - `agent`: Custom agent to be used. Default is to create the skills in the `src/skills.ts` file.
 - `hideInitLogMessage`: hide the init log message with messagekit logo and stuff
