+++ packages/message-kit/src/index.ts
@@ -4,5 +4,6 @@ export * from "./lib/xmtp.js";
 export * from "./helpers/types.js";
 export * from "./helpers/gpt.js";
 export * from "./helpers/resolver.js";
+export * from "./helpers/cdp.js";
 export { Client as V2Client } from "@xmtp/xmtp-js";
 export { Client as V3Client } from "@xmtp/node-sdk";
